+++ .pre-commit-config.yaml
@@ -43,11 +43,11 @@ repos:
   #       exclude: "tests"
 
   # markdown linting
-  - repo: https://github.com/igorshubovych/markdownlint-cli
-    rev: v0.33.0
-    hooks:
-      - id: markdownlint
-        args: [--config=.markdownlint.yaml]
+  # - repo: https://github.com/igorshubovych/markdownlint-cli
+  #   rev: v0.33.0
+  #   hooks:
+  #     - id: markdownlint
+  #       args: [--config=.markdownlint.yaml]
 
   # Git conflict marker
   - repo: https://github.com/jumanjihouse/pre-commit-hooks
