+++ datumaro/components/hl_ops/__init__.py
@@ -26,34 +26,34 @@ __all__ = ["HLOps"]
 class HLOps:
     """High-level dataset operations for Python API."""
 
-    @staticmethod
-    def compare(
-        src_dataset: IDataset,
-        tgt_dataset: IDataset,
-        level: str = "high",
-        *,
-        env: Optional[Environment] = None,
-        **kwargs,
-    ) -> IDataset:
-        """
-        Applies some function to dataset items.
+    # @staticmethod
+    # def compare(
+    #     src_dataset: IDataset,
+    #     tgt_dataset: IDataset,
+    #     level: str = "high",
+    #     *,
+    #     env: Optional[Environment] = None,
+    #     **kwargs,
+    # ) -> IDataset:
+    #     """
+    #     Applies some function to dataset items.
 
-        Results are computed lazily, if the transform supports this.
+    #     Results are computed lazily, if the transform supports this.
 
-        Args:
-            dataset: The dataset to be transformed
-            method: The transformation to be applied to the dataset.
-                If a string is passed, it is treated as a plugin name,
-                which is searched for in the environment
-                set by the 'env' argument
-            env: A plugin collection. If not set, the built-in plugins are used
-            **kwargs: Parameters for the transformation
+    #     Args:
+    #         dataset: The dataset to be transformed
+    #         method: The transformation to be applied to the dataset.
+    #             If a string is passed, it is treated as a plugin name,
+    #             which is searched for in the environment
+    #             set by the 'env' argument
+    #         env: A plugin collection. If not set, the built-in plugins are used
+    #         **kwargs: Parameters for the transformation
 
-        Returns: a wrapper around the input dataset
-        """
+    #     Returns: a wrapper around the input dataset
+    #     """
 
-        comparator = Comparator()
-        return comparator.compare_datasets(src_dataset, tgt_dataset, level=level)
+    #     comparator = Comparator()
+    #     return comparator.compare_datasets(src_dataset, tgt_dataset, level=level)
 
     @staticmethod
     def transform(
